Fix group command (exit) of example scripts

+++ example/update_intact.sh
@@ -9,7 +9,7 @@ WGET_LOG=wget.log
 
 XML_DIR=pmid
 
-wget -c -m http://$DB_FTP/$ZIP_FILE -o $WGET_LOG || ( cat $WGET_LOG; exit 1 )
+wget -c -m http://$DB_FTP/$ZIP_FILE -o $WGET_LOG || cat $WGET_LOG && exit 1
 
 grep 'not retrieving' $WGET_LOG > /dev/null
 
@@ -20,7 +20,7 @@ if [ $? = 0 ] && [ -d $XML_DIR ] ; then
 
 fi
 
-grep 'No such file' $WGET_LOG > /dev/null && ( cat $WGET_LOG; exit 1 )
+grep 'No such file' $WGET_LOG > /dev/null && cat $WGET_LOG && exit 1
 
 rm -rf $XML_DIR
